move EntityPermissions to models project
